+++ src/Context.php
@@ -37,7 +37,6 @@ interface Context
      * Returns the step used by this context.
      *
      * If no cash rounding is involved, this must return 1.
-     * This value is used by money allocation methods that do not go through the applyTo() method.
      *
      * @return positive-int
      *